Add RoR Home controller
Create controller with index action method to provide entry point
for the application.
Set root to the home#index in routes.
Respond only in HTML format.
Respond with status code 500 when view template is not found.
As a result, the controller should respond with status code 500
when json format is requested.

Add RSpec tests for the following controllers:
  - HomeController

Add the following feature specs:
  - User visits home page

Improvements:
  - Update main page HTML layout

Integrate the following tools:
  - Bootstrap
  - Capybara